# docs/swagger.yaml
basePath: /api/v1
definitions:
  messages.Error:
    properties:
      ERROR: {}
    type: object
  messages.ListResponse:
    properties:
      messages:
        items:
          $ref: '#/definitions/messages.Message'
        type: array
    type: object
  messages.Message:
    properties:
      external_message_id:
        type: string
      sent_at:
        type: string
    type: object
  service.Error:
    properties:
      ERROR: {}
    type: object
  service.ServiceStatus:
    enum:
    - STOPPED
    - WORKING
    type: string
    x-enum-varnames:
    - ServiceStatusStopped
    - ServiceStatusWorking
  service.ToggleRequest:
    type: object
  service.ToggleResponse:
    properties:
      SERVICE_STATUS:
        $ref: '#/definitions/service.ServiceStatus'
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample messaging worker.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Sample Messaging Worker
  version: "1.0"
paths:
  /messages/list:
    get:
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/messages.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/messages.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/messages.Error'
  /services/toggle:
    post:
      consumes:
      - application/json
      description: toggle worker status
      parameters:
      - description: Toggle Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.ToggleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/service.ToggleResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/service.Error'
      summary: toggle worker status
      tags:
      - Worker Service
swagger: "2.0"